# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Book a Seat – Workspace & Meeting Room Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meeting-rooms:
    get:
      operationId: listMeetingRooms
      tags: [meeting-rooms]
      summary: List meeting rooms
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: ["string", "null"]
        - name: capacity
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of meeting rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MeetingRoom"
    post:
      operationId: createMeetingRoom
      tags: [meeting-rooms]
      summary: Create a meeting room (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingRoomInput"
      responses:
        "201":
          description: Created meeting room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingRoom"
  /meeting-rooms/{id}:
    get:
      operationId: getMeetingRoom
      tags: [meeting-rooms]
      summary: Get meeting room by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingRoom"
    patch:
      operationId: updateMeetingRoom
      tags: [meeting-rooms]
      summary: Update meeting room (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingRoomUpdate"
      responses:
        "200":
          description: Updated meeting room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingRoom"
    delete:
      operationId: deleteMeetingRoom
      tags: [meeting-rooms]
      summary: Delete meeting room (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # MEETING ROOM BOOKINGS
  /meeting-room-bookings:
    get:
      operationId: listMeetingRoomBookings
      tags: [meeting-room-bookings]
      summary: List meeting room bookings
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: query
          schema:
            type: ["integer", "null"]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Paginated meeting room bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMeetingRoomBookings"
    post:
      operationId: createMeetingRoomBooking
      tags: [meeting-room-bookings]
      summary: Book a meeting room (team lead only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingRoomBookingInput"
      responses:
        "201":
          description: Created meeting room booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingRoomBooking"
  /meeting-room-bookings/{id}:
    get:
      operationId: getMeetingRoomBooking
      tags: [meeting-room-bookings]
      summary: Get meeting room booking by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting room booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingRoomBooking"
    patch:
      operationId: updateMeetingRoomBooking
      tags: [meeting-room-bookings]
      summary: Update meeting room booking (team lead only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingRoomBookingUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingRoomBooking"
  /meeting-room-bookings/{id}/cancel:
    post:
      operationId: cancelMeetingRoomBooking
      tags: [meeting-room-bookings]
      summary: Cancel a meeting room booking (team lead only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingRoomBooking"

  # NOTIFICATIONS
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ANALYTICS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [employee, team_lead, admin]
        department:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    MeetingRoom:
      type: object
      required: [id, name, capacity, floor, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        capacity:
          type: integer
        floor:
          type: string
        status:
          type: string
          enum: [available, maintenance]
        facilities:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    MeetingRoomInput:
      type: object
      required: [name, capacity, floor]
      properties:
        name:
          type: string
        capacity:
          type: integer
        floor:
          type: string
        facilities:
          type: string
        description:
          type: string
    MeetingRoomUpdate:
      type: object
      properties:
        name:
          type: string
        capacity:
          type: integer
        floor:
          type: string
        facilities:
          type: string
        description:
          type: string
        status:
          type: string
    MeetingRoomBooking:
      type: object
      required: [id, roomId, userId, title, startTime, endTime, status, createdAt]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        startTime:
          type: string
        endTime:
          type: string
        status:
          type: string
          enum: [confirmed, cancelled]
        attendees:
          type: ["string", "null"]
        createdAt:
          type: string
        room:
          $ref: "#/components/schemas/MeetingRoom"
        user:
          $ref: "#/components/schemas/User"
    MeetingRoomBookingInput:
      type: object
      required: [roomId, title, startTime, endTime]
      properties:
        roomId:
          type: integer
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        attendees:
          type: string
    MeetingRoomBookingUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        attendees:
          type: string
    PaginatedMeetingRoomBookings:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MeetingRoomBooking"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Notification:
      type: object
      required: [id, userId, type, message, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [booking_confirmed, booking_cancelled, booking_expired, maintenance_alert, meeting_reminder, check_in_reminder]
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
...